+++ src/main/java/booking_stay/booking_stay/couponevent/domain/enums/CouponEventStatus.java
@@ -1,0 +1,16 @@
+package booking_stay.booking_stay.couponevent.domain.enums;
+
+import lombok.Getter;
+
+@Getter
+public enum CouponEventStatus {
+    READY("RD"),
+    DO("DO"),
+    FINISH("FN");
+
+    private final String code;
+
+    CouponEventStatus(String code) {
+        this.code = code;
+    }
+}
