--- src/main/java/booking_stay/booking_stay/couponeventV2/domain/enums/CouponEventStatus.java
@@ -1,16 +1,0 @@
-package booking_stay.booking_stay.couponeventV2.domain.enums;
-
-import lombok.Getter;
-
-@Getter
-public enum CouponEventStatus {
-    READY("RD"),
-    DO("DO"),
-    FINISH("FN");
-
-    private final String code;
-
-    CouponEventStatus(String code) {
-        this.code = code;
-    }
-}
