--- src/main/java/booking_stay/booking_stay/couponeventV2/dto/CouponEventUpdateRequestDto.java
@@ -1,25 +1,0 @@
-package booking_stay.booking_stay.couponeventV2.dto;
-
-import booking_stay.booking_stay.couponeventV2.domain.enums.CouponEventStatus;
-import lombok.Builder;
-import lombok.Getter;
-
-import java.time.LocalDateTime;
-
-@Getter
-public class CouponEventUpdateRequestDto {
-    private String eventName;
-    private int maxQuantity;
-    private CouponEventStatus status;
-    private LocalDateTime startTime;
-    private LocalDateTime endTime;
-
-    @Builder
-    public CouponEventUpdateRequestDto(String eventName, int maxQuantity, CouponEventStatus status, LocalDateTime startTime, LocalDateTime endTime) {
-        this.eventName = eventName;
-        this.maxQuantity = maxQuantity;
-        this.status = status;
-        this.startTime = startTime;
-        this.endTime = endTime;
-    }
-}
