--- src/main/java/booking_stay/booking_stay/couponeventV2/service/CouponEventRedisService.java
@@ -1,14 +1,0 @@
-package booking_stay.booking_stay.couponeventV2.service;
-
-import booking_stay.booking_stay.couponeventV2.domain.entity.CouponEventRequest;
-
-public interface CouponEventRedisService {
-
-    String couponEventProducer(CouponEventRequest request);
-
-    void couponEventConsumer();
-
-    String trunOnCouponEvent(Long id);
-
-    String resetCouponEventCount(Long id);
-}
