+++ src/test/java/booking_stay/booking_stay/couponevent/service/impl/CouponEventRedisServiceImplTest.java
@@ -1,0 +1,7 @@
+package booking_stay.booking_stay.couponevent.service.impl;
+
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CouponEventRedisServiceImplTest {
+}
